backend/pyproject.toml
[project]
name = "career-agent"
version = "0.1.0"
description = "CareerAgent — Multi-Agent LLM Job Application Assistant"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi==0.135.2",
    "uvicorn[standard]==0.42.0",
    "python-multipart>=0.0.18",
    # Config & validation
    "pydantic-settings==2.13.1",
    # Database
    "supabase==2.28.3",
    "psycopg[binary]>=3.2.0",
    # LangChain / LangGraph suite (March 2026 latest)
    "langchain-core==1.2.21",
    "langchain-openai==1.1.11",
    "langchain-community==0.4.1",
    "langgraph==1.1.3",
    "langgraph-checkpoint-postgres==3.0.5",
    # External APIs
    "tavily-python==0.7.23",
    # CV parsing
    "pymupdf>=1.25.0",
    # HTTP
    "httpx>=0.28.0",
    "langgraph-cli[inmem]>=0.4.19",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.9.0",
    "black>=25.0.0",
    "mypy>=1.14.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
